Commit of snapshot 20140809
Several chip/device vendors leverage a subset
of this codebase as one of their .so files.

So, we're going to provide at least this line in the sand
as a commit/tag so people can download this as a release
from GitHub instead of www.soft-switch.org which is DEAD
right now.