Commit of Snapshot 20180108
- This consists of the last bits of work that
was downloadable at this time via the Wayback
Machine for this library.  It is unknown how
well this compiles at this time.
- The intention for going forward for this on
this branch is to ensure it works as intended
as this is used by a lot of people.  This will
be followed with a likely move to CMake or Meson
for this so it can more cleanly build in all
contexts, including cross-compile.